extract tabular table and mobile filters into context-driven components
